+++ backends/yaml/sync.js
@@ -22,7 +22,7 @@ module.exports = {
         });
     },
 
-    postMissing: function(lng, ns, key, defaultValue, callback) {
+    saveMissing: function(lng, ns, key, defaultValue, callback) {
         console.log('not implemented');
         callback('not implemented');
     },