+++ sample/server.js
@@ -59,7 +59,8 @@ i18n.init({
     ns: { namespaces: ['ns.common', 'ns.special'], defaultNs: 'ns.special'},
     resSetPath: 'locales/__lng__/new.__ns__.json',
     saveMissing: true,
-    debug: true
+    debug: true,
+    sendMissingTo: 'all'
 });
 
 // Configuration